Goals loading and setting is now functional and some cleanup. Basic implementation of all features except ai is now completed. Improvements in ui are all thats left to be added

+++ digital_twin_server/src/main/java/com/nhlstenden/groep3/digital_twin/DTO/BlockDTO.java
@@ -15,22 +15,22 @@ public class BlockDTO {
 
     private Polygon geometry;
 
-    private float height;
+    private double height;
 
-    private float area;
+    private double area;
 
-    private float volume;
+    private double volume;
 
-    private float calculated_cost;
+    private double calculated_cost;
 
-    private float calculated_yield;
+    private double calculated_yield;
 
     private int calculated_residents;
 
 
     public BlockDTO(){}
 
-    public BlockDTO(Long id, BlockType blockType, Long mapId, Polygon geometry, float height, float area, float volume, float calculated_cost, float calculated_yield, int calculated_residents) {
+    public BlockDTO(Long id, BlockType blockType, Long mapId, Polygon geometry, double height, double area, double volume, double calculated_cost, double calculated_yield, int calculated_residents) {
         this.id = id;
         this.blockType = blockType;
         this.mapId = mapId;
@@ -75,43 +75,43 @@ public class BlockDTO {
         this.geometry = geometry;
     }
 
-    public float getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public void setHeight(float height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 
-    public float getArea() {
+    public double getArea() {
         return area;
     }
 
-    public void setArea(float area) {
+    public void setArea(double area) {
         this.area = area;
     }
 
-    public float getVolume() {
+    public double getVolume() {
         return volume;
     }
 
-    public void setVolume(float volume) {
+    public void setVolume(double volume) {
         this.volume = volume;
     }
 
-    public float getCalculated_cost() {
+    public double getCalculated_cost() {
         return calculated_cost;
     }
 
-    public void setCalculated_cost(float calculated_cost) {
+    public void setCalculated_cost(double calculated_cost) {
         this.calculated_cost = calculated_cost;
     }
 
-    public float getCalculated_yield() {
+    public double getCalculated_yield() {
         return calculated_yield;
     }
 
-    public void setCalculated_yield(float calculated_yield) {
+    public void setCalculated_yield(double calculated_yield) {
         calculated_yield = calculated_yield;
     }
